Make ArchivedEntityBanner a context component (#60939)
